SDI-266 Package rename to reflect the scope of these services
