+++ src/main/java/uk/gov/justice/hmpps/prison/service/receiveandtransfer/BookNumberGenerationService.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.hmpps.prison.service.transfer
+package uk.gov.justice.hmpps.prison.service.receiveandtransfer
 
 import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Service
