+++ src/test/java/uk/gov/justice/hmpps/prison/service/receiveandtransfer/BedAssignmentTransferServiceTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.hmpps.prison.service.transfer
+package uk.gov.justice.hmpps.prison.service.receiveandtransfer
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
